requirements.txt
# Core
jupyter
ipykernel
python-dotenv

# Data processing
pandas
numpy
pyarrow

# Database
sqlalchemy
psycopg2-binary

# Data collection
requests
feedparser
beautifulsoup4
praw
google-api-python-client

# Utilities
tenacity
tqdm
pydantic
faker
pyyaml
rich
sqlparse

# Visualization
matplotlib
plotly
openpyxl

# Infrastructure
minio
prefect
boto3

# NLP & Annotation (E2/E3)
spacy
yake

# Development
ruff
nbformat


